refactor: Replace under package name 'global/entity' #1
# Conflicts:
#	src/main/java/com/profect/delivery/global/entity/Menu.java
