+++ src/main/java/com/profect/delivery/domain/menu/repository/MenuRepository.java
@@ -1,6 +1,6 @@
 package com.profect.delivery.domain.menu.repository;
 
-import com.profect.delivery.domain.menu.entity.Menu;
+import com.profect.delivery.global.entity.Menu;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
