+++ src/main/java/com/profect/delivery/global/entity/Menu.java
@@ -1,5 +1,6 @@
 package com.profect.delivery.global.entity;
 
+import com.fasterxml.jackson.annotation.JsonProperty;
 import jakarta.persistence.*;
 import lombok.*;
 
@@ -9,12 +10,12 @@ import java.util.UUID;
 @Table(name = "p_menus")
 @IdClass(MenuId.class)
 @Getter
+@AllArgsConstructor(access = AccessLevel.PRIVATE)
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Builder
 @Setter
-@NoArgsConstructor
-@AllArgsConstructor
 public class Menu {
-
-    @Id
+    @Id @GeneratedValue
     @Column(name = "menu_id", nullable = false)
     private UUID menuId;
 
@@ -39,24 +40,32 @@ public class Menu {
     @Column(name = "is_public", nullable = false)
     private boolean isPublic;
 
-    @Column(name = "created_at", nullable = false)
+    @Column(nullable = false, updatable = false)
     private LocalDateTime createdAt;
-
-    @Column(name = "created_by", length = 10, nullable = false)
+    @Column(nullable = false, length = 10)
     private String createdBy;
 
-    @Column(name = "updated_at")
     private LocalDateTime updatedAt;
-
-    @Column(name = "updated_by", length = 10)
+    @Column(length = 10)
     private String updatedBy;
 
-    @Column(name = "deleted_at")
-    private LocalDateTime deletedAt;
-
-    @Column(name = "deleted_by", length = 10)
+    @Column private LocalDateTime deletedAt;
+    @Column(length = 10)
     private String deletedBy;
+    private String deletedRs;
 
-    @Column(name = "deleted_rs", length = 100)
-    private String deletedReason;
+    public void changeName(String name)               { this.menuName = name; }
+    public void changePrice(Integer price)            { this.menuPrice = price; }
+    public void changeCategory(UUID categoryId)       { this.menuCategoryId = categoryId; }
+    public void changeDescription(String description) { this.menuDescription = description; }
+    public void changeIsPublic(Boolean isPublic)      { this.isPublic = isPublic; }
+
+    public void auditUpdate(String username) {
+        this.updatedAt = LocalDateTime.now();
+        this.updatedBy = username;
+    }
+    public void auditDelete(String username) {
+        this.deletedAt = LocalDateTime.now();
+        this.deletedBy = username;
+    }
 }
